// src/generated/rustplus.ts
export interface AppMessage {
  response?: AppResponse | undefined;
  broadcast?: AppBroadcast | undefined;
}
// ...
export interface AppBroadcast {
  teamChanged?: AppTeamChanged | undefined;
  teamMessage?: AppNewTeamMessage | undefined;
  entityChanged?: AppEntityChanged | undefined;
  clanChanged?: AppClanChanged | undefined;
  clanMessage?: AppNewClanMessage | undefined;
  cameraRays?: AppCameraRays | undefined;
}
// ...
export interface AppCameraInfo {
  width: number;
  height: number;
  nearPlane: number;
  farPlane: number;
  controlFlags: number;
}

export interface AppCameraRays {
  verticalFov: number;
  sampleOffset: number;
  rayData: Uint8Array;
  distance: number;
  entities: AppCameraRays_Entity[];
}

export enum AppCameraRays_EntityType {
  Tree = 1,
  Player = 2,
}

export interface AppCameraRays_Entity {
  entityId: number;
  type: AppCameraRays_EntityType;
  position?: Vector3 | undefined;
  rotation?: Vector3 | undefined;
  size?: Vector3 | undefined;
  name?: string | undefined;
}
// ...
export const AppMessage: MessageFns<AppMessage> = {
  encode(message: AppMessage, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.response !== undefined) {
      AppResponse.encode(message.response, writer.uint32(10).fork()).join();
    }
    if (message.broadcast !== undefined) {
      AppBroadcast.encode(message.broadcast, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AppMessage {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAppMessage();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.response = AppResponse.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.broadcast = AppBroadcast.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<AppMessage>, I>>(base?: I): AppMessage {
    return AppMessage.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AppMessage>, I>>(object: I): AppMessage {
    const message = createBaseAppMessage();
    message.response = (object.response !== undefined && object.response !== null)
      ? AppResponse.fromPartial(object.response)
      : undefined;
    message.broadcast = (object.broadcast !== undefined && object.broadcast !== null)
      ? AppBroadcast.fromPartial(object.broadcast)
      : undefined;
    return message;
  },
};
// ...
export const AppCameraInfo: MessageFns<AppCameraInfo> = {
  encode(message: AppCameraInfo, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.width !== 0) {
      writer.uint32(8).int32(message.width);
    }
    if (message.height !== 0) {
      writer.uint32(16).int32(message.height);
    }
    if (message.nearPlane !== 0) {
      writer.uint32(29).float(message.nearPlane);
    }
    if (message.farPlane !== 0) {
      writer.uint32(37).float(message.farPlane);
    }
    if (message.controlFlags !== 0) {
      writer.uint32(40).int32(message.controlFlags);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AppCameraInfo {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAppCameraInfo();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.width = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.height = reader.int32();
          continue;
        }
        case 3: {
          if (tag !== 29) {
            break;
          }

          message.nearPlane = reader.float();
          continue;
        }
        case 4: {
          if (tag !== 37) {
            break;
          }

          message.farPlane = reader.float();
          continue;
        }
        case 5: {
          if (tag !== 40) {
            break;
          }

          message.controlFlags = reader.int32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<AppCameraInfo>, I>>(base?: I): AppCameraInfo {
    return AppCameraInfo.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AppCameraInfo>, I>>(object: I): AppCameraInfo {
    const message = createBaseAppCameraInfo();
    message.width = object.width ?? 0;
    message.height = object.height ?? 0;
    message.nearPlane = object.nearPlane ?? 0;
    message.farPlane = object.farPlane ?? 0;
    message.controlFlags = object.controlFlags ?? 0;
    return message;
  },
};

function createBaseAppCameraRays(): AppCameraRays {
  return { verticalFov: 0, sampleOffset: 0, rayData: new Uint8Array(0), distance: 0, entities: [] };
}

export const AppCameraRays: MessageFns<AppCameraRays> = {
  encode(message: AppCameraRays, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.verticalFov !== 0) {
      writer.uint32(13).float(message.verticalFov);
    }
    if (message.sampleOffset !== 0) {
      writer.uint32(16).int32(message.sampleOffset);
    }
    if (message.rayData.length !== 0) {
      writer.uint32(26).bytes(message.rayData);
    }
    if (message.distance !== 0) {
      writer.uint32(37).float(message.distance);
    }
    for (const v of message.entities) {
      AppCameraRays_Entity.encode(v!, writer.uint32(42).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AppCameraRays {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAppCameraRays();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 13) {
            break;
          }

          message.verticalFov = reader.float();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.sampleOffset = reader.int32();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.rayData = reader.bytes();
          continue;
        }
        case 4: {
          if (tag !== 37) {
            break;
          }

          message.distance = reader.float();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.entities.push(AppCameraRays_Entity.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<AppCameraRays>, I>>(base?: I): AppCameraRays {
    return AppCameraRays.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AppCameraRays>, I>>(object: I): AppCameraRays {
    const message = createBaseAppCameraRays();
    message.verticalFov = object.verticalFov ?? 0;
    message.sampleOffset = object.sampleOffset ?? 0;
    message.rayData = object.rayData ?? new Uint8Array(0);
    message.distance = object.distance ?? 0;
    message.entities = object.entities?.map((e) => AppCameraRays_Entity.fromPartial(e)) || [];
    return message;
  },
};

function createBaseAppCameraRays_Entity(): AppCameraRays_Entity {
  return { entityId: 0, type: 1, position: undefined, rotation: undefined, size: undefined, name: "" };
}

export const AppCameraRays_Entity: MessageFns<AppCameraRays_Entity> = {
  encode(message: AppCameraRays_Entity, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.entityId !== 0) {
      writer.uint32(8).uint32(message.entityId);
    }
    if (message.type !== 1) {
      writer.uint32(16).int32(message.type);
    }
    if (message.position !== undefined) {
      Vector3.encode(message.position, writer.uint32(26).fork()).join();
    }
    if (message.rotation !== undefined) {
      Vector3.encode(message.rotation, writer.uint32(34).fork()).join();
    }
    if (message.size !== undefined) {
      Vector3.encode(message.size, writer.uint32(42).fork()).join();
    }
    if (message.name !== undefined && message.name !== "") {
      writer.uint32(50).string(message.name);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AppCameraRays_Entity {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAppCameraRays_Entity();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.entityId = reader.uint32();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.type = reader.int32() as any;
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.position = Vector3.decode(reader, reader.uint32());
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.rotation = Vector3.decode(reader, reader.uint32());
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.size = Vector3.decode(reader, reader.uint32());
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.name = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<AppCameraRays_Entity>, I>>(base?: I): AppCameraRays_Entity {
    return AppCameraRays_Entity.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AppCameraRays_Entity>, I>>(object: I): AppCameraRays_Entity {
    const message = createBaseAppCameraRays_Entity();
    message.entityId = object.entityId ?? 0;
    message.type = object.type ?? 1;
    message.position = (object.position !== undefined && object.position !== null)
      ? Vector3.fromPartial(object.position)
      : undefined;
    message.rotation = (object.rotation !== undefined && object.rotation !== null)
      ? Vector3.fromPartial(object.rotation)
      : undefined;
    message.size = (object.size !== undefined && object.size !== null) ? Vector3.fromPartial(object.size) : undefined;
    message.name = object.name ?? "";
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}